Refactor: separating blob and local saving logic
