+++ src/pages/api/uploadBlob.js
@@ -1,15 +1,31 @@
-import { put } from '@vercel/blob';
+//import { put, list } from '@vercel/blob';
+import { requestRainRadar } from '@/lib/fmiQueryData';
+import { fetchRadarImages } from '@/lib/radarImageUtils/blobHandling';
 
-export default async function handler(request, response) {
-    const blob = await put(request.query.filename, request, {
-        access: 'public',
-    });
+export default async function handler(req, res) {
+    if (req.method === 'POST') {
+        const { timestamps } = req.body;
 
-    return response.status(200).json({ url: blob.url });
+        if (!timestamps || timestamps.length === 0) {
+            return res.status(400).json({ error: 'No timestamps provided' });
+        }
+
+        try {
+            const urls = timestamps.map((time) => {
+                const config = requestRainRadar(time);
+                return `${config.url}?${new URLSearchParams(config.params).toString()}`;
+            });
+
+            const imagePaths = await fetchRadarImages(urls, timestamps);
+
+            const result = imagePaths;
+
+            res.status(200).json({ imagePaths: result });
+        } catch (error) {
+            res.status(500).json({ error: 'Failed to process radar images' });
+            throw new Error(error);
+        }
+    } else {
+        res.status(405).json({ error: 'Method not allowed' });
+    }
 }
-
-export const config = {
-    api: {
-        bodyParser: false,
-    },
-};
